+++ test/contractsv2/BridgeV2GasTokens.test.ts
@@ -4,22 +4,10 @@ import { setBalance } from '@nomicfoundation/hardhat-network-helpers';
 import { MTBridge, mtBridgeUtils } from '@0xpolygonhermez/zkevm-commonjs';
 import { computeWrappedTokenProxyAddress } from './helpers/helpers-sovereign-bridge';
 import { ERC20PermitMock, PolygonZkEVMGlobalExitRoot, AgglayerBridge, TokenWrapped } from '../../typechain-types';
+import { calculateGlobalExitRoot, computeGlobalIndex } from '../../src/utils';
 
 const MerkleTreeBridge = MTBridge;
 const { verifyMerkleProof, getLeafValue } = mtBridgeUtils;
-
-function calculateGlobalExitRoot(mainnetExitRoot: any, rollupExitRoot: any) {
-    return ethers.solidityPackedKeccak256(['bytes32', 'bytes32'], [mainnetExitRoot, rollupExitRoot]);
-}
-// eslint-disable-next-line @typescript-eslint/naming-convention
-const _GLOBAL_INDEX_MAINNET_FLAG = 2n ** 64n;
-
-function computeGlobalIndex(indexLocal: any, indexRollup: any, isMainnet: boolean) {
-    if (isMainnet === true) {
-        return BigInt(indexLocal) + _GLOBAL_INDEX_MAINNET_FLAG;
-    }
-    return BigInt(indexLocal) + BigInt(indexRollup) * 2n ** 32n;
-}
 
 describe('PolygonZkEVMBridge Gas tokens tests', () => {
     upgrades.silenceWarnings();
