+++ test/contractsv2/UpgradeToPP.test.ts
@@ -18,6 +18,7 @@ import {
 import { encodeInitializeBytesLegacy } from '../../src/utils-common-aggchain';
 import { VerifierType, computeRandomBytes } from '../../src/pessimistic-utils';
 import { AL_MULTISIG_ROLE } from '../../src/constants';
+import { calculateGlobalExitRoot, calculateGlobalExitRootLeaf, calculateAccInputHashetrog } from '../../src/utils';
 
 const MerkleTreeBridge = MTBridge;
 const { getLeafValue } = mtBridgeUtils;
@@ -359,7 +360,7 @@ describe('Upgradeable to PPV2 or ALGateway', () => {
             forcedBlockHashL1: ethers.ZeroHash,
         } as PolygonValidiumEtrog.ValidiumBatchDataStruct;
 
-        const expectedAccInputHash = calculateAccInputHashEtrog(
+        const expectedAccInputHash = calculateAccInputHashetrog(
             await validiumContract.lastAccInputHash(),
             hashedData,
             await polygonZkEVMGlobalExitRoot.getRoot(),
@@ -616,7 +617,7 @@ describe('Upgradeable to PPV2 or ALGateway', () => {
             forcedBlockHashL1: ethers.ZeroHash,
         } as PolygonValidiumEtrog.ValidiumBatchDataStruct;
 
-        const expectedAccInputHash = calculateAccInputHashEtrog(
+        const expectedAccInputHash = calculateAccInputHashetrog(
             await validiumContract.lastAccInputHash(),
             hashedData,
             await polygonZkEVMGlobalExitRoot.getRoot(),
@@ -905,7 +906,7 @@ describe('Upgradeable to PPV2 or ALGateway', () => {
             '0x', // empty metadata
         );
 
-        const expectedAccInputHash = calculateAccInputHashEtrog(
+        const expectedAccInputHash = calculateAccInputHashetrog(
             ethers.ZeroHash,
             ethers.keccak256(transaction),
             await polygonZkEVMGlobalExitRoot.getLastGlobalExitRoot(),
@@ -951,7 +952,7 @@ describe('Upgradeable to PPV2 or ALGateway', () => {
 
         const rootSC = await polygonZkEVMGlobalExitRoot.getRoot();
 
-        const expectedAccInputHash2 = calculateAccInputHashEtrog(
+        const expectedAccInputHash2 = calculateAccInputHashetrog(
             expectedAccInputHash,
             ethers.keccak256(l2txData),
             rootSC,
@@ -1226,7 +1227,7 @@ describe('Upgradeable to PPV2 or ALGateway', () => {
             forcedBlockHashL1: ethers.ZeroHash,
         } as PolygonValidiumEtrog.ValidiumBatchDataStruct;
 
-        const expectedAccInputHash = calculateAccInputHashEtrog(
+        const expectedAccInputHash = calculateAccInputHashetrog(
             await validiumContract.lastAccInputHash(),
             hashedData,
             await polygonZkEVMGlobalExitRoot.getRoot(),
@@ -1384,40 +1385,4 @@ describe('Upgradeable to PPV2 or ALGateway', () => {
 
         expect(await rollupManagerContract.isRollupMigrating(newCreatedRollupID)).to.be.equal(false);
     });
-
-    /**
-     * Compute accumulateInputHash = Keccak256(oldAccInputHash, batchHashData, l1InfoTreeRoot, timestamp, seqAddress)
-     * @param {String} oldAccInputHash - old accumulateInputHash
-     * @param {String} batchHashData - Batch hash data
-     * @param {String} globalExitRoot - Global Exit Root
-     * @param {Number} timestamp - Block timestamp
-     * @param {String} sequencerAddress - Sequencer address
-     * @returns {String} - accumulateInputHash in hex encoding
-     */
-    function calculateAccInputHashEtrog(
-        oldAccInputHash: any,
-        batchHashData: any,
-        l1InfoTreeRoot: any,
-        timestamp: any,
-        sequencerAddress: any,
-        forcedBlockHash: any,
-    ) {
-        const hashKeccak = ethers.solidityPackedKeccak256(
-            ['bytes32', 'bytes32', 'bytes32', 'uint64', 'address', 'bytes32'],
-            [oldAccInputHash, batchHashData, l1InfoTreeRoot, timestamp, sequencerAddress, forcedBlockHash],
-        );
-
-        return hashKeccak;
-    }
-
-    function calculateGlobalExitRoot(mainnetExitRoot: any, rollupExitRoot: any) {
-        return ethers.solidityPackedKeccak256(['bytes32', 'bytes32'], [mainnetExitRoot, rollupExitRoot]);
-    }
-
-    function calculateGlobalExitRootLeaf(newGlobalExitRoot: any, lastBlockHash: any, timestamp: any) {
-        return ethers.solidityPackedKeccak256(
-            ['bytes32', 'bytes32', 'uint64'],
-            [newGlobalExitRoot, lastBlockHash, timestamp],
-        );
-    }
 });
