+++ test/contractsv2/claimCompressor/ClaimCompressor.test.ts
@@ -4,6 +4,7 @@ import { ethers, upgrades } from 'hardhat';
 import { takeSnapshot } from '@nomicfoundation/hardhat-network-helpers';
 import { processorUtils, MTBridge, mtBridgeUtils } from '@0xpolygonhermez/zkevm-commonjs';
 import { PolygonZkEVMGlobalExitRoot, AgglayerBridge, ClaimCompressor, BridgeReceiverMock } from '../../typechain-types';
+import { calculateGlobalExitRoot, computeGlobalIndex } from '../../src/utils';
 
 const MerkleTreeBridge = MTBridge;
 const { getLeafValue } = mtBridgeUtils;
@@ -21,19 +22,6 @@ function calculateCallDataCost(calldataBytes: string): number {
     }
 
     return totalCost;
-}
-
-function calculateGlobalExitRoot(mainnetExitRoot: any, rollupExitRoot: any) {
-    return ethers.solidityPackedKeccak256(['bytes32', 'bytes32'], [mainnetExitRoot, rollupExitRoot]);
-}
-// eslint-disable-next-line @typescript-eslint/naming-convention
-const _GLOBAL_INDEX_MAINNET_FLAG = 2n ** 64n;
-
-function computeGlobalIndex(indexLocal: any, indexRollup: any, isMainnet: boolean) {
-    if (isMainnet === true) {
-        return BigInt(indexLocal) + _GLOBAL_INDEX_MAINNET_FLAG;
-    }
-    return BigInt(indexLocal) + BigInt(indexRollup) * 2n ** 32n;
 }
 
 describe('Claim Compressor Contract', () => {
