+++ test/contractsv2/helpers/helpers-sovereign-bridge.ts
@@ -1,23 +1,9 @@
 import { ethers } from 'hardhat';
 import { MTBridge, mtBridgeUtils } from '@0xpolygonhermez/zkevm-commonjs';
+import { calculateGlobalExitRoot, computeGlobalIndex } from '../../../src/utils';
 
 const { getLeafValue } = mtBridgeUtils;
 const MerkleTreeBridge = MTBridge;
-
-// Constants
-// eslint-disable-next-line @typescript-eslint/naming-convention
-const _GLOBAL_INDEX_MAINNET_FLAG = 2n ** 64n;
-
-export function computeGlobalIndex(indexLocal: any, indexRollup: any, isMainnet: boolean) {
-    if (isMainnet === true) {
-        return BigInt(indexLocal) + _GLOBAL_INDEX_MAINNET_FLAG;
-    }
-    return BigInt(indexLocal) + BigInt(indexRollup) * 2n ** 32n;
-}
-
-export function calculateGlobalExitRoot(mainnetExitRoot: any, rollupExitRoot: any) {
-    return ethers.solidityPackedKeccak256(['bytes32', 'bytes32'], [mainnetExitRoot, rollupExitRoot]);
-}
 
 export async function computeWrappedTokenProxyAddress(
     networkId: any,
